Add mixins for background images
For the mixins to work, it was necessary to pass data to sass using functions. And this was not possible with the old _sass_ plugin, so I had to replace it. And so that _postcss-url_ does not break paths in `@supports`, I had to put it after _sass_, but at the same time _postcss-import_ became unnecessary.

If there are already functions, then the values of breakpoints for viewports are also now transferred from the project data.